+++ network.tf
@@ -10,13 +10,6 @@ resource "azurerm_virtual_network" "vnet" {
   address_space       = var.vnet_cidrs
 }
 
-resource "azurerm_subnet" "bastion" {
-  name                 = "AzureBastionSubnet" # it needs to be exactly this name
-  resource_group_name  = azurerm_resource_group.db_rg.name
-  virtual_network_name = azurerm_virtual_network.vnet.name
-  address_prefixes     = var.bastion_subnet_cidrs
-}
-
 resource "azurerm_subnet" "pe" {
   name                 = "pe-subnetnet"
   resource_group_name  = azurerm_resource_group.db_rg.name
@@ -66,18 +59,6 @@ resource "azurerm_network_security_group" "postgres_db_nsg" {
   resource_group_name = azurerm_resource_group.db_rg.name
 
   security_rule {
-    name                       = "allow-bastion-ingress"
-    priority                   = 100
-    direction                  = "Inbound"
-    access                     = "Allow"
-    protocol                   = "Tcp"
-    source_port_range          = "*"
-    destination_port_range     = "5432"
-    source_address_prefix      = azurerm_subnet.bastion.address_prefixes[0]
-    destination_address_prefix = azurerm_subnet.db.address_prefixes[0]
-  }
-
-  security_rule {
     name                       = "allow-app-ingress"
     priority                   = 101
     direction                  = "Inbound"
@@ -105,4 +86,27 @@ resource "azurerm_network_security_group" "postgres_db_nsg" {
 resource "azurerm_subnet_network_security_group_association" "postgres_nsg_assoc" {
   subnet_id                 = azurerm_subnet.db.id
   network_security_group_id = azurerm_network_security_group.postgres_db_nsg.id
+}
+
+resource "azurerm_network_security_group" "app_nsg" {
+  name                = "appSubnetNSG"
+  location            = var.location
+  resource_group_name = azurerm_resource_group.db_rg.name
+
+  security_rule {
+    name                       = "allow-bastion-ssh"
+    priority                   = 100
+    direction                  = "Inbound"
+    access                     = "Allow"
+    protocol                   = "Tcp"
+    source_port_range          = "*"
+    destination_port_range     = "22"
+    source_address_prefix      = "VirtualNetwork"
+    destination_address_prefix = azurerm_subnet.app.address_prefixes[0]
+  }
+}
+
+resource "azurerm_subnet_network_security_group_association" "app_nsg_assoc" {
+  subnet_id                 = azurerm_subnet.app.id
+  network_security_group_id = azurerm_network_security_group.app_nsg.id
 }